refactor(intelligence): move SourceEvidence to evidence.models
